pyproject.toml
[project]
name = "TAPPER"
dynamic = ["version"]
description = "HARDWARIO TAPPER - an MQTT NFC reader"
readme = "README.md"
authors = [
    { name = "Martin", email = "81148675+Martian-0007@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "adafruit-blinka>=8.56.0",
    "adafruit-circuitpython-pn532>=2.4.3",
    "click>=8.1.8",
    "gpiozero>=2.0.1",
    "lgpio",
    "loguru>=0.7.3",
    "paho-mqtt>=2.1.0",
    "psutil>=7.0.0",
    "pyyaml>=6.0.2",
    "uvloop>=0.21.0",
]

[project.scripts]
tapper = "tapper:cli.cli"

[project.optional-dependencies]
lgpio = []
logtail = [
    "logtail-python>=0.3.3",
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[dependency-groups]
dev = [
    "hatch>=1.14.0",
    "ruff>=0.11.2",
]

[tool.hatch.build.hooks.version]
path = "src/tapper/_version.py"
